Debugg
Debugging I seem to get the clientID to show now, had to check and debugg the query and figured out why! Its because for get all clients its * in Select * From Clients this is wrong since it then allows for redundant columns which overwrite previous ones. Specifying which columns were needed exactly fixed this